Added basic camera

+++ source/weapon/RangedWeaponBase.h
@@ -24,7 +24,7 @@ protected:
     float attackCD = 0.0;
     float baseAttackCD = 0.2f;
     float comboPointsIncrease = 20.0f;
-    float damage = 1.0f;
+    int damage = 1.0f;
     float soundStrength = 0.6f;
 
     Sprite sprite;
